split footer into a component
