+++ src/App.js
@@ -19,7 +19,7 @@ import {
   Heading,
   Container,
   Hero, } from "react-bulma-components/full";
-import { CenteredColumn, EmailForm, TileBox } from './components/componentIndex';
+import { CenteredColumn, EmailForm, TileBox, RhonenFooter } from './components/componentIndex';
 import './helpers/fontawesome';
 
 class App extends Component {
@@ -81,15 +81,7 @@ class App extends Component {
           </Icon>
         </span>
         </Section>
-        <Footer>
-          <Container>
-            <Level>
-              <Level.Item>
-                <p>Made with <strike>love</strike> <span><FontAwesomeIcon icon={["fab", "react"]} size="2x" transform="down-3" /></span> by <strong>Joel Rhine</strong> aka <em>rhonen</em> <span role="img" aria-label="praying hands emoji">🙏</span></p>
-              </Level.Item>
-            </Level>
-          </Container>
-        </Footer>
+        <RhonenFooter />
       </div>      
     );
   }
