--- src/components/WelcomeModal.js
@@ -1,27 +1,0 @@
-import React, { Component } from 'react';
-import { 
-    OpenModal,
-    Image,
-    Media,
-    Button,
-    Content,
-    Modal,
-    Card,
-    Icon,
-    Footer,
-    Level,
-    Tag,
-    Tile,
-    Section,
-    Heading,
-    Container,
-    Hero, } from "react-bulma-components/full";
-
-//TODO: Refactor footer in App.js here...
-export class RhonenFooter extends Component {
-    render() {
-        return (
-            <div></div>
-        )
-    }
-}
