+++ src/components/componentIndex.js
@@ -1,4 +1,4 @@
 export * from './CenteredColumn';
 export * from './EmailForm';
 export * from './Tiles';
-export * from './WelcomeModal';
+export * from './RhonenFooter';
